+++ lib/napa/generators.rb
@@ -1,3 +1,4 @@
 require 'napa/generators/scaffold_generator'
 require 'napa/generators/api_generator'
 require 'napa/generators/migration_generator'
+require 'napa/generators/readme_generator'