+++ spec/generators/readme_generator_spec.rb
@@ -1,0 +1,35 @@
+require 'spec_helper'
+require 'napa/generators/readme_generator'
+require 'napa/cli'
+
+describe Napa::Generators::ReadmeGenerator do
+  let(:test_readme_directory) { 'spec/tmp' }
+
+  before do
+    described_class.any_instance.stub(:output_directory) { test_readme_directory }
+    Napa::CLI::Base.new.generate("readme")
+  end
+
+  after do
+    FileUtils.rm_rf(test_readme_directory)
+  end
+
+  describe 'README' do
+    it 'creates a README in the current directory' do
+      expected_readme_file = File.join(test_readme_directory, 'README.md')
+      readme = File.read(expected_readme_file)
+
+      expect(readme).to match /# #{Napa::Identity.name}/
+    end
+  end
+
+  describe 'spec' do
+    it 'creates a README spec' do
+      expected_spec_file = File.join(test_readme_directory, 'spec/docs/readme_spec.rb')
+      spec_code = File.read(expected_spec_file)
+
+      expect(spec_code).to match(/describe \'README\'/)
+    end
+  end
+
+end